just refreshing

+++ src/LoopsAndConditionals/LoopsAndConditionals12.java
@@ -3,7 +3,7 @@ package LoopsAndConditionals;
 import javax.swing.*;
 import java.util.Scanner;
 
-//Using
+//Using Math.PI, Math.pow, System.out.printf, /n, Integer.parseInt, Double.parseInt
 
 public class LoopsAndConditionals12 {
 
@@ -15,21 +15,21 @@ public class LoopsAndConditionals12 {
 
         switch (figura) {
             case 1 :
-                double side = Integer.parseInt(JOptionPane.showInputDialog("Plese enter the lenght of the side"));
+                double side = Double.parseDouble(JOptionPane.showInputDialog("Plese enter the lenght of the side"));
                 System.out.println("the square area is: " + Math.pow(side, 2));
                 break;
             case 2 :
-                double base = Integer.parseInt(JOptionPane.showInputDialog("Plese enter base lenght"));
-                double height = Integer.parseInt(JOptionPane.showInputDialog("Plese enter height"));
+                double base = Double.parseDouble(JOptionPane.showInputDialog("Plese enter base lenght"));
+                double height = Double.parseDouble(JOptionPane.showInputDialog("Plese enter height"));
                 System.out.println("The rectangle area is: " + (base * height));
                 break;
             case 3:
-                base = Integer.parseInt(JOptionPane.showInputDialog("Plese enter base lenght"));
-                height = Integer.parseInt(JOptionPane.showInputDialog("Plese enter height"));
+                base = Double.parseDouble(JOptionPane.showInputDialog("Plese enter base lenght"));
+                height = Double.parseDouble(JOptionPane.showInputDialog("Plese enter height"));
                 System.out.println("The triangle area is: " + (base*height)/2);
                 break;
             case 4:
-                double radio = Integer.parseInt(JOptionPane.showInputDialog("Plese enter the lenght of the radio"));
+                double radio = Double.parseDouble(JOptionPane.showInputDialog("Plese enter the lenght of the radio"));
                 System.out.println("The circle area is: ");
                 System.out.printf("%1.2f", Math.PI*Math.pow(radio,2));
                 break;
